Update transit-gateway-quotas.md
Changed from connection to tunnel as this is per tunnel and not connection. 

Note: This is correctly stated in VPN FAQ: https://aws.amazon.com/vpn/faqs/
"Q: What is the approximate maximum throughput of a Site-to-Site VPN connection?

A: Each AWS Site-to-Site VPN connection has two tunnels and each tunnel supports a maximum throughput of up to 1.25 Gbps. If your VPN connection is to a Virtual Private Gateway, aggregated throughput limits would apply."