Update transit-gateway-quotas.md
Added line #17 for clarity. The way original line #16 come across is confusing that you can have multiple TGW attachments from a single TGW to the same VPC, which is not true. The original sentence line #16 actually mean one VPC can have up to 5 attachments to 5 separate TGWs. Fee free to re-word the original line #16 if it can be made clear vs adding an extra note (line #17) as long as it read correctly. Thanks