Update tgw-peering.md
Added line #11 for clarity and consistency. Reference: https://aws.amazon.com/about-aws/whats-new/2020/04/aws-transit-gateway-now-supports-inter-region-peering-in-11-additional-regions/. Note that line #11 show in the public docs but missing here in the Github copy.

The Github version is out of sync from line #10 - #13 from the public docs. This make proposing any update the docs very challenging as you see on thing in the public docs, but then when you look at the Github version for the same it's missing. I hope this can be fixed.